feat: backup old state and set new ltredux main state
